backend/requirements.txt
fastapi
uvicorn[standard]
langdetect
rank-bm25
FlagEmbedding
pydantic
python-dotenv
httpx
pdfplumber
pytesseract
Pillow
